upd: reorg/condense
